+++ docs/CHANGELOG.md
@@ -70,6 +70,7 @@ Version 1.5.0 (Trillium): Not yet released
 - Fix a one frame flash to the default pose each time an animation loops on a mesh being simulated in the level editor
 - Update Weather Region bounds in viewport live when values are changed
 - Fix crash when a bone, tag, or prop point is queried on a character whose mesh file contains no submeshes, or when such a character's weapon or textures are preloaded for rendering
+- Fix items not rotating correctly when out of view or when viewed through a Projection Camera
 
 [@is-this-c](https://github.com/is-this-c)
 - Let `Caps Lock` capitalize